Merging transaction is not fully auth yet due to new features being added.

+++ src/main/java/com/ase/restservice/service/AccountService.java
@@ -8,11 +8,13 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import static com.ase.restservice.ApplicationSecurity.getUsernameOfClientLogged;
+
 /**
  * Service for Account operations.
  */
 @Service
-public class AccountService implements AccountServiceI {
+public class AccountService implements com.ase.restservice.service.AccountServiceI {
 
   @Autowired
   private AccountRepository accountRepository;
@@ -35,6 +37,9 @@ public class AccountService implements AccountServiceI {
       );
     } catch (AccountNotFoundException e) {
       // reach here means account does not already exist in database
+      //clientID placement must be handled by system.
+      String clientID = getUsernameOfClientLogged();
+      account.setClientId(clientID);
       return accountRepository.save(account);
     }
   }
@@ -46,6 +51,7 @@ public class AccountService implements AccountServiceI {
    * @return Updated account
    */
   public Account updateAccount(Account account) throws AccountNotFoundException {
+    //TODO might cause an security risk as users shouldnt be able to change the client
     try {
       String accountId = account.getAccountId();
       Account dbAccount = this.getAccountById(accountId);
@@ -63,7 +69,13 @@ public class AccountService implements AccountServiceI {
   public void deleteAccountById(String accountId) throws AccountNotFoundException {
     try {
       Account dbAccount = this.getAccountById(accountId);
-      accountRepository.deleteById(accountId);
+      //TODO check if i can delete it
+
+      String clientID = getUsernameOfClientLogged();
+
+      if (clientID.equals(dbAccount.getClientId())) {
+        accountRepository.deleteById(accountId);
+      }
     } catch (AccountNotFoundException e) {
       throw new AccountNotFoundException(e);
     }
@@ -84,7 +96,9 @@ public class AccountService implements AccountServiceI {
   }
 
   /**
-   * List all accounts.
+   * THIS SHOULD NOT BE USED DUE TO SECURITY.
+   * THIS HAS TO BE HERE FOR COMPILER
+   * Lists all accounts.
    *
    * @return list of accounts
    */
@@ -107,6 +121,14 @@ public class AccountService implements AccountServiceI {
     account.setBalance(account.getBalance() + amount);
     final Account updatedAccount = this.updateAccount(account);
     return updatedAccount;
+  }
+  /**
+   * List all accounts that logged in client owns.
+   *
+   * @return list of accounts
+   */
+  public List<Account> listAllAccountsOfClient(String clientId) {
+    return accountRepository.findAllAccountByClient(clientId);
   }
 
 }
